Added DefaultQOS option for an association.

+++ src/common/print_fields.c
@@ -214,8 +214,9 @@ extern void print_fields_uint32(print_field_t *field, uint32_t value, int last)
 extern void print_fields_uint64(print_field_t *field, uint64_t value, int last)
 {
 	int abs_len = abs(field->len);
+
 	/* (value == unset)  || (value == cleared) */
-	if((value == NO_VAL) || (value == INFINITE)) {
+	if((value == (uint64_t)NO_VAL) || (value == (uint64_t)INFINITE)) {
 		if(print_fields_parsable_print
 		   == PRINT_FIELDS_PARSABLE_NO_ENDING
 		   && last)
